render product

+++ Library/src/main/java/com/ecommerce/library/dto/ProductDto.java
@@ -12,8 +12,8 @@ public class ProductDto {
     private Long id;
     private String name;
     private String description;
-    private double costPrice;
-    private double salePrice;
+    private int costPrice;
+    private int salePrice;
     private int currentQuantity;
     private Category category;
     private String image;
